Clarification: Decay Matrix shows relative Decay not Attenuation.
Attenuation is the reciproce of a gain  factor. But here, the Decay Matrix shows the decay to be applied from the relation of original and recorded signal. So if y has longer decay, that behavior van be observed directly in the Decay Matrix for each frequency bin. Because the Decay Matrix is convolved with the impulse response, which will be deconvoluted from a third signal z, it can't be called attenuation anymore.